Refactor login and signup form elements
I changed many things in the code. First, I removed the summary and replaced it with an h2. Second, I added a class to the div and an ID to the input. Third, I changed the type of the buttons from sign up and log in to submit. Finally, I added the link of the CSS file. I hope you like it